feat: Use shallow clone in installation instructions
This commit updates the installation instructions in the documentation to use a shallow clone when cloning the repository. 

By adding `--depth 1` to the `git clone` command, we instruct Git to only download the latest commit, rather than the full commit history. This significantly reduces the amount of data transferred during the cloning process, leading to faster downloads and lower storage requirements for the user. 

The documentation has been updated to reflect this change and explain the purpose of the `--depth 1` option.